refactor: move readCsv and validateInput into parsing layer
readCsv.js and validateInput.js belong to parsing, not application
orchestration. src/parsing/ now has no imports from src/application/,
making the parsing module fully self-contained.

Updated imports in parseInput.js (./readCsv.js, ./validateInput.js),
PdfToCsvAdapter.test.js, and moved validateInput.test.js to
src/parsing/__tests__/.
